+++ WebFile.BlazorServer/Pages/Index.razor.cs
@@ -30,12 +30,12 @@ public sealed partial class Index
 
         if (user == null)
             return;
-        
+
         await using var context = await DbFactory.CreateDbContextAsync();
         var userModel = await context.Users.Include(x => x.Files)
             .FirstOrDefaultAsync(x => x.Equals(user));
         if (userModel == null) return;
-        
+
         IsAuthenticated = true;
         User = user;
         FolderModels = userModel.GetFolder();
@@ -57,12 +57,13 @@ public sealed partial class Index
         if (!Directory.Exists(filePath)) Directory.CreateDirectory(filePath);
 
         var fileName = $"{User.UserName}/{Convert.ToBase64String(
-            System.Text.Encoding.Default.GetBytes($"{DateTime.Now:s}{Guid.NewGuid().ToString()}"))}/{arg.OriginFileName}";
+            System.Text.Encoding.Default.GetBytes($"{DateTime.Now:s}{Guid.NewGuid()}"))}/{arg.OriginFileName}";
         var saveFilePath = Path.Combine(filePath, fileName);
 
-        if (await arg.SaveToFileAsync(saveFilePath, 1012 * 1024, new CancellationTokenSource().Token))
+        if (await arg.SaveToFileAsync(saveFilePath))
         {
-            var pwd = $"{User.UserName}{DateTime.Now:s}{Guid.NewGuid().ToString()}{arg.OriginFileName}".HashEncryption().Replace("/", "-");
+            var pwd = $"{User.UserName}{DateTime.Now:s}{Guid.NewGuid()}{arg.OriginFileName}".HashEncryption()
+                .Replace("/", "-");
             user.Files.Add(new FileModel() { Path = fileName, Id = pwd });
             await context.SaveChangesAsync();
             await ToastService.Success("上传文件成功");
@@ -72,31 +73,48 @@ public sealed partial class Index
         }
 
         await ToastService.Error("上传文件", $"保存文件失败 {arg.OriginFileName}");
+        var fileInfo = new FileInfo(saveFilePath);
+        if (fileInfo.Exists) fileInfo.Delete();
+        var dirInfo = new DirectoryInfo(Path.GetDirectoryName(saveFilePath)!);
+        if (dirInfo.Exists) dirInfo.Delete();
     }
 
-    private async Task OnDelete(FolderModel model)
+    private async Task OnDelete(IFile model)
     {
         await using var context = await DbFactory.CreateDbContextAsync();
         var userModel = await context.Users.Include(x => x.Files)
             .FirstOrDefaultAsync(x => x.Equals(User));
         if (userModel == null) return;
-        var file = await context.FileModel.Include(x => x.Owner).FirstOrDefaultAsync(x => x.Id == model.Id);
-        if (file == null) return;
-        userModel.Files.Remove(file);
-        var fileInfo = new FileInfo($@"{WebHostEnvironment.WebRootPath}\UserFiles\{file.Path}");
-        fileInfo.Delete();
+        var item = userModel.Files.FirstOrDefault(x => x.Id == model.Id);
+        if (item == null) return;
+
+        if (!item.IsFolder)
+        {
+            DeleteFile(item);
+        }
+        else
+        {
+            var list = userModel.Files.Where(x => x.Url == model.Path).ToList();
+            foreach (var t in list.Where(t => t.Id != model.Id))
+            {
+                DeleteFile(t);
+                userModel.Files.Remove(t);
+            }
+        }
+
+        userModel.Files.Remove(item);
         await context.SaveChangesAsync();
         FolderModels = userModel.GetFolder();
         StateHasChanged();
     }
 
-    private async Task OnDownload(FolderModel model)
+    private async Task OnDownload(IFile model)
     {
         await DownloadService.DownloadFromStreamAsync(Path.GetFileName(model.Path),
             model.GetStream(WebHostEnvironment.WebRootPath));
     }
 
-    private Color GetIconColor(FolderModel model)
+    private Color GetIconColor(IFile model)
     {
         var i = FolderModels.FindIndex(fileModel => fileModel.Id == model.Id);
         i %= 11;
@@ -131,7 +149,14 @@ public sealed partial class Index
                     return false;
                 }
 
-                var pwd = $"{User.UserName}{DateTime.Now:s}{Guid.NewGuid().ToString()}{path}".HashEncryption().Replace("/", "-");
+                if (user.Files.Any(x => Path.GetFileName(x.Path) == path))
+                {
+                    await ToastService.Error("添加文件夹", "有重名文件夹");
+                    return false;
+                }
+
+                var pwd = $"{User.UserName}{DateTime.Now:s}{Guid.NewGuid().ToString()}{path}".HashEncryption()
+                    .Replace("/", "-");
                 user.Files.Add(new FileModel() { Path = path, Id = pwd, IsFolder = true });
                 await context.SaveChangesAsync();
                 await ToastService.Success("添加文件夹成功");
@@ -144,10 +169,19 @@ public sealed partial class Index
         await DialogService.ShowEditDialog(option);
     }
 
-    private Task OnOpen(FolderModel model)
+    private Task OnOpen(IFile model)
     {
         NavigationManager.NavigateTo(model.ToWebUrl());
         return Task.CompletedTask;
+    }
+
+    private void DeleteFile(IFile file)
+    {
+        var path = $@"{WebHostEnvironment.WebRootPath}\UserFiles\{file.Path}";
+        var fileInfo = new FileInfo(path);
+        fileInfo.Delete();
+        var dirInfo = new DirectoryInfo(Path.GetDirectoryName(path)!);
+        dirInfo.Delete();
     }
 }
